Xcos: be compatible with Scilab 5.5.2 zcos files
To ease transition from Xcos 5 to Xcos 6, the ZCOS file format is
compatible. To have the same rendering on both version, the encoding of
some values (style rotation) have to be set in a special manner.

Port geometry is still discarded as it is no more present on the Port
model. The user have to manually resize each block to enforce a port
layouting on Scilab 5 after openning a Scilab 6 saved file.

+++ scilab/modules/graph/src/java/org/scilab/modules/graph/shape/SvgShape.java
@@ -76,8 +76,8 @@ public class SvgShape extends mxLabelShape {
      */
     @Override
     public Rectangle getImageBounds(mxGraphics2DCanvas canvas, mxCellState state) {
-        final double rotation = mxUtils.getDouble(state.getStyle(),
-                                mxConstants.STYLE_ROTATION, 0);
+        final int rotation = (int) mxUtils.getDouble(state.getStyle(),
+                             mxConstants.STYLE_ROTATION, 0);
         final AffineTransform transform = new AffineTransform();
         transform.rotate(Math.toRadians(rotation), state.getCenterX(),
                          state.getCenterY());
